Use FGT to determine a lowrank affinity matrix
Instead of building a giant affinity matrix G to determine the lowrank
eigenvectors, we implicit generate the affinity matrix using figtree.
For now, we're letting figtree run fast and loose with
FIGTREE_EVAL_AUTO, but if we need our affinity matrix to be exact we
could switch it to FIGTREE_EVAL_DIRECT. That will be slower, but the
memory footprint will still be lower than the previous setup.