Automatically determine numeig by default
This is clumsy for small datasets, since it produces crappy results. But
then again, you shouldn't be using nonrigid_lowrank for small datasets
anyways.

Could be some space for future work sanity-checking the size of datasets
passed into various algorithms, but for now we'll just let users shoot
themselves in the foot.