Allow configuration of sigma2
Sigma2 is provided to a registration as an unscaled value, which is then
scaled down for the algorithm proper. Sigma2 should be a bit more than
the maximum "wiggle" expected in the dataset - e.g. if you know your
data moved about five meters to the side, sigma2 should be a bit more
than 5 meters.

At least I think that's the way it should work. Worth investigating
more.