Add CPDConfig.cmake and friends
This creates a new exported target (cpd-lib) that can be used in
downstream projects. We've created #19 to hold a tutorial that describes
the usage of the cmake exported target.

Right now our version string doesn't *exactly* match what cmake wants --
cmake's version math doesn't handle strings or dashes sanely. However, I
think it'd be over-engineering to worry about that now, since (a) we
don't know how much the cmake version logic will be used and (b) I don't
know how likely it is that we'll ever get out of that wonderful 0.x
release land where I can be looser with breaking stuff. Something to be
aware of, though.

Fixes #15