Fix broken examples on Travis
Docopt needs boost to work w/ gcc 4.8, which is on travis. We don't
actually build the docopt example on Travis, but we previously were
bringing in docopt to both examples. This patch adds mechanisms to only
include docopt in the artificial-displacement example, and builds
against boost if we have the buggy gcc version.

This was a good test of `cmake_parse_arguments` and including docopt for
old versions — I have a feeling we might use this knowledge in the
future.

Fixes #99